Clean repo - remove vendor/core/files from tracking

+++ web/modules/contrib/tamper/src/Plugin/Tamper/TruncateText.php
@@ -4,8 +4,8 @@ namespace Drupal\tamper\Plugin\Tamper;
 
 use Drupal\Component\Utility\Unicode;
 use Drupal\Core\Form\FormStateInterface;
-use Drupal\tamper\TamperableItemInterface;
 use Drupal\tamper\TamperBase;
+use Drupal\tamper\TamperableItemInterface;
 
 /**
  * Plugin implementation of the Truncate Text plugin.
@@ -14,7 +14,8 @@ use Drupal\tamper\TamperBase;
  *   id = "truncate_text",
  *   label = @Translation("Truncate"),
  *   description = @Translation("Truncate"),
- *   category = "Text"
+ *   category = @Translation("Text"),
+ *   itemUsage = "ignored"
  * )
  */
 class TruncateText extends TamperBase {
@@ -74,7 +75,12 @@ class TruncateText extends TamperBase {
   /**
    * {@inheritdoc}
    */
-  public function tamper($data, TamperableItemInterface $item = NULL) {
+  public function tamper($data, ?TamperableItemInterface $item = NULL) {
+    // Don't process empty or null values.
+    if (is_null($data) || $data === '') {
+      return $data;
+    }
+
     return Unicode::truncate(
       $data,
       $this->getSetting(self::SETTING_NUM_CHAR),
